+++ NEWS.md
@@ -30,6 +30,7 @@ but has been expanded to `wflow_git_commit()`, `wflow_publish()`, and
     * The argument `latest` is no longer mutually-exclusive with `files`. If the
     most recently modified HTML is different than those specified by `files`,
     they will all be opened for viewing.
+    * S3 print method
 
 [included-files]: https://rmarkdown.rstudio.com/rmarkdown_websites.html#included-files